Program 7
Accept passenger name and total miles travelled
a. If miles > 10,000, award 10 frequent flyer points to the passenger and display
b. If miles > 20,000, award 20 frequent flyer points to the passenger and display
c. If miles > 50,000, award 30 frequent flyer points to the passenger and display
d. If miles > 100,000, award 50 frequent flyer points to the passenger and display